backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
python-socketio==5.11.3
eventlet==0.36.1

opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4

groq==0.11.0
python-dotenv==1.0.1
httpx==0.27.0